reorganisation of files and methods (#390)
